+++ common/src/main/java/games/alejandrocoria/imagetitles/mixin/MixinGui.java
@@ -1,8 +1,8 @@
 package games.alejandrocoria.imagetitles.mixin;
 
-import com.llamalad7.mixinextras.sugar.Local;
 import games.alejandrocoria.imagetitles.ImageTitles;
 import net.minecraft.client.DeltaTracker;
+import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.Gui;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
@@ -11,6 +11,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 @Mixin(Gui.class)
 public class MixinGui {
@@ -24,8 +25,9 @@ public class MixinGui {
                     target="Lnet/minecraft/util/Mth;clamp(III)I",
                     shift = At.Shift.AFTER),
             method = "renderTitle",
+            locals = LocalCapture.CAPTURE_FAILSOFT,
             cancellable = true)
-    private void renderTitle(GuiGraphics guiGraphics, DeltaTracker deltaTracker, CallbackInfo ci, @Local int i) {
+    private void renderTitle(GuiGraphics guiGraphics, DeltaTracker deltaTracker, CallbackInfo ci, Font font, float f, int i) {
         boolean imageRendered = ImageTitles.renderImage(guiGraphics, i);
 
         if (imageRendered) {