+++ forge/src/main/java/games/alejandrocoria/imagetitles/client/ImageTitlesClientForge.java
@@ -1,13 +1,11 @@
 package games.alejandrocoria.imagetitles.client;
 
 import games.alejandrocoria.imagetitles.ImageTitles;
-import net.minecraft.client.Minecraft;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.server.packs.resources.SimplePreparableReloadListener;
 import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
 import net.minecraftforge.event.TickEvent;
-import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;