+++ doc/changelog.md
@@ -2,6 +2,8 @@ Metadata has been moved to .mcmeta files with a different format (https://github
 
 The previous format is still supported in this version, but probably not in the next.
  
+
+* [Forge] Fixed: Mixin error.
  
 **Requirements:**
 * ${loadername}: ${loaderversion}